% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDavidTilingData.R
\name{getDavidTilingData}
\alias{getDavidTilingData}
\title{Create the data set davidTiling.}
\usage{
getDavidTilingData()
}
\value{
An ExpressionSet object containing the davidTiling data.
}
\description{
In version older than 1.53.0, this dataset was provided as a pre-built 
ExpressionSet object, available via `data(davidTiling)`. 
With this function, the dataset is built on-the-fly by:
- downloading the raw CEL files from ArrayExpress (E-TABM-14)
- reading the CEL files into an ExpressionSet object
}
\examples{
\donttest{
getDavidTilingData()
}

}
\references{
David L, Huber W, Granovskaia M, et al. A high-resolution map of
transcription in the yeast genome. Proceedings of the National Academy of
Sciences of the United States of America. 2006 Apr;103(14):5320-5325.
\doi{10.1073/pnas.0601091103}. PMID: 16569694; PMCID: PMC1414796.
}
\seealso{
\url{https://www.ebi.ac.uk/arrayexpress/files/E-TABM-14/}
}
