\name{data:davidTiling}
\alias{data:davidTiling}
\alias{davidTiling}
\non_function{}
\title{Dataset of class eSet with the raw 'CEL file' intensities}
\usage{data("davidTiling")}

\format{Intensity data for 8 arrays. The
\code{\link[Biobase]{phenoData}} slot contains the file names
and the nucleic acid type.}

\description{
This dataset is no longer provided as a pre-built ExpressionSet object,
but can be built on-the-fly by calling \code{\link{getDavidTilingData}}.
}

\source{
Lior David and Lars Steinmetz, both from the Stanford
Genome Center. Lars Steinmetz is also at EMBL Heidelberg.
}
\author{W. Huber \email{huber@ebi.ac.uk}}
\keyword{datasets}
